data-copilot/requirements.txt
Flask
Werkzeug
crewai
google-auth
google-cloud-bigquery
langchain-google-vertexai
vertexai
